New Function: PairwiseAligner for build_structures.py
New Function using PairwiseAligner for build_structures.py, instead of pairwise2.

And it uses spatial info of AAs to Identify suspicious “misplaced” chunks near gaps and moves them to a more appropriate place in the alignment.

But the code block is commented out in the build/management/commands/build_structures.py